requirements.txt
# Minimum Django version
Django==1.9
django-audit-log==0.7.0

# Helps with backwards compatibility before Python 3.5
simplejson==3.8.1

# Test requirements
tox==2.3.1
pytest-django==2.9.1
pytest==2.8.5
pytest-cov==2.2.0
flake8==2.5.1
factory_boy==2.6.1

# wheel for PyPI installs
wheel==0.26.0

# Sphinx for documentation builds
Sphinx==1.3.3
